// src/app/base-modules/m-tooltable/models/tooltable.model.ts
export enum EToolTableAction {
    INSERT = 'INSERT',
    UPDATE = 'UPDATE',
    CLONE = 'CLONE',
    DELETE = 'DELETE',
    EXPORT = 'EXPORT',
    UPDATE_STATUS = 'UPDATE_STATUS',
    VIEW = 'VIEW',
    MONITOR_WEB = 'MONITOR_WEB',
    MONITOR_SOCIAL = 'MONITOR_SOCIAL',
    NOTE = 'NOTE',
    ROLE = 'ROLE',
}